requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
bidict==0.23.1
billiard==4.2.2
blinker==1.9.0
brotli==1.2.0
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
ConfigArgParse==1.7.1
Django==5.2.8
django-prometheus==2.4.1
djangorestframework==3.16.1
dnspython==2.8.0
drf-spectacular==0.29.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.9.1
geventhttpclient==2.3.5
greenlet==3.2.4
h11==0.16.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
locust==2.42.2
locust-cloud==1.28.1
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==25.0
platformdirs==4.5.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psutil==7.1.3
Pygments==2.19.2
pymongo==4.15.3
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-socketio==5.14.3
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.4
rpds-py==0.28.0
simple-websocket==1.1.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==6.1
zope.interface==8.0.1
